with 32: 2.20 s run, 6.60 s total

+++ covasim/people.py
@@ -73,7 +73,7 @@ class People(cvb.BasePeople):
 
         prognoses = pars['prognoses']
         age_cutoffs = prognoses['age_cutoffs']
-        inds = np.fromiter((find_cutoff(age_cutoffs, this_age) for this_age in self.age), dtype=np.int64, count=len(self))
+        inds = np.fromiter((find_cutoff(age_cutoffs, this_age) for this_age in self.age), dtype=np.int32, count=len(self))
         self.symp_prob[:]    = self.pars['rel_symp_prob']   * prognoses['symp_probs'][inds]
         self.severe_prob[:]  = pars['rel_severe_prob'] * prognoses['severe_probs'][inds]
         self.crit_prob[:]    = pars['rel_crit_prob']   * prognoses['crit_probs'][inds]
