+++ covasim/population.py
@@ -105,7 +105,7 @@ def make_randpop(sim, age_data=None, sex_ratio=0.5, microstructure=False):
     age_data_range = age_data_max - age_data_min
     age_data_prob = age_data[:,2]
     age_data_prob /= age_data_prob.sum() # Ensure it sums to 1
-    age_bins = cvu.multinomial(np.array(age_data_prob, dtype=np.float64), np.int64(pop_size)) # Choose age bins
+    age_bins = cvu.multinomial(np.array(age_data_prob, dtype=np.float32), np.int32(pop_size)) # Choose age bins
     ages = age_data_min[age_bins] + age_data_range[age_bins]*np.random.random(pop_size) # Uniformly distribute within this age bin
 
     # Store output; data duplicated as per-person and list-like formats for convenience
@@ -188,7 +188,7 @@ def make_microstructured_contacts(pop_size, contacts):
             n_remaining -= this_cluster
 
         for key in contacts_dict.keys():
-            contacts_list[key][layer_name] = np.array(list(contacts_dict[key]), dtype=np.int64)
+            contacts_list[key][layer_name] = np.array(list(contacts_dict[key]), dtype=np.int32)
 
     return contacts_list, contact_keys
 
@@ -255,7 +255,7 @@ def make_synthpop(sim):
             int_contacts[new_key] = []
             for uid in uid_contacts[key]:
                 int_contacts[new_key].append(uid_mapping[uid])
-            int_contacts[new_key] = np.array(int_contacts[new_key], dtype=np.int64)
+            int_contacts[new_key] = np.array(int_contacts[new_key], dtype=np.int32)
         contacts.append(int_contacts)
 
     popdict = {}
